+++ src/frontend/apps/desk/src/features/mail-domains/access-management/components/index.ts
@@ -1,4 +1,3 @@
 export * from './AccessesList';
 export * from './ChooseRole';
-export * from './ModalRole';
 export * from './ModalDelete';