+++ src/frontend/apps/desk/src/features/mail-domains/mailboxes/components/panel/MailBoxesListView.tsx
@@ -1,4 +1,4 @@
-import { DataGrid, SortModel } from '@openfun/cunningham-react';
+import { DataGrid, SortModel } from '@gouvfr-lasuite/cunningham-react';
 import type { InfiniteData } from '@tanstack/react-query';
 import { useEffect, useMemo, useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
@@ -129,6 +129,14 @@ export function MailBoxesListView({
     <div>
       {error && <TextErrors causes={error.cause ?? []} />}
 
+      {!mailDomain.count_mailboxes && (
+        <Box $align="center" $margin={{ top: 'base', bottom: 'base' }}>
+          <Text $size="small">
+            {t('No mail box was created with this mail domain.')}
+          </Text>
+        </Box>
+      )}
+
       {filteredMailboxes && filteredMailboxes.length ? (
         <>
           <DataGrid
@@ -141,7 +149,7 @@ export function MailBoxesListView({
                 renderCell: ({ row }) => (
                   <Text
                     $weight={row.isCurrentUser ? '600' : '400'}
-                    $theme={row.isCurrentUser ? 'primary' : 'greyscale'}
+                    $theme={row.isCurrentUser ? 'primary' : 'gray'}
                   >
                     {row.email}
                   </Text>
@@ -154,7 +162,7 @@ export function MailBoxesListView({
                 renderCell: ({ row }) => (
                   <Text
                     $weight="500"
-                    $theme="greyscale"
+                    $theme="gray"
                     $css="text-transform: capitalize;"
                   >
                     {row.name}
@@ -186,10 +194,11 @@ export function MailBoxesListView({
             ]}
             isLoading={isLoading}
           />
+
           {isFetchingNextPage && <div>{t('Loading more...')}</div>}
         </>
       ) : null}
-      <div ref={loadMoreRef} style={{ height: 32 }} />
+      <div ref={loadMoreRef} style={{ height: 0 }} />
     </div>
   );
 }